requirements.txt
wheel
lxml
python-docx
docx2pdf
docxcompose

